requirements.txt
annotated-types==0.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==23.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
dataclasses==0.6
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
google==3.0.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jwt==1.3.1
MarkupSafe==2.1.5
mistune==3.0.2
packaging==24.0
psycopg2==2.9.9
py-dotenv==0.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyMySQL==1.1.0
python-engineio==4.9.0
python-socketio==5.11.2
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
simple-websocket==1.0.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
wsproto==1.2.0
uvicorn==0.27.0
gunicorn==21.2.0
Flask[async]
cryptography==42.0.7
pycryptodome==3.20.0
PyPDF2==3.0.1
rsa==4.9
pdfkit==1.0.0
qrcode==7.4.2
opencv-python==4.9.0.80
